sstables: extract entry_descriptor into a separate header
